+++ certbot-dns-cloudxns/docs/api/dns_cloudxns.rst
@@ -1,5 +1,5 @@
-:mod:`certbot_dns_cloudxns.dns_cloudxns`
+:mod:`certbot_dns_cloudxns._internal.dns_cloudxns`
 ----------------------------------------
 
-.. automodule:: certbot_dns_cloudxns.dns_cloudxns
+.. automodule:: certbot_dns_cloudxns._internal.dns_cloudxns
    :members:
